Use cached addr for disconnected devices
The "PeripheralId" isn't something we can pass
over to C code (because it has private fields we can't access). So we are using the address.
Unfortunately, in the "DeviceDisconnected" event, the peripheral is gone, so we can't query it again.

So we are keeping a hash map of all the device addresses we've seen and using that to map the disconnect message.

That's super frustrating, because the PerhiperalId _is_ the address under the covers in many implementations, but we can't access it.